Acceptance tests using Behave, WebTest and Nose work!
The hard work for this is done in django_nose which provides the
appropriate WebTest app, which we can simply re-use in the behave
environment.py.

They need to be run using:

    behave tests/features

Note, you need to use:

    behave --no-capture tests/features

for PDB to work.

First feature is not erroring

Simple homepage tests works

Extra scenario added